tests/test_album_repository.py
from lib.album_repository import AlbumRepository
from lib.album import Album

"""
When I call #all
I get all the albums in the albums table
"""

def test_all(db_connection):
    db_connection.seed("seeds/record_store.sql")
    repository = AlbumRepository(db_connection)
    assert repository.all() == [
        Album(1, "hello bob", 2008, 1)
    ]


"""
When I call #create
I create an album in the database
And I can see it back in #all
"""

def test_create(db_connection):
    db_connection.seed("seeds/record_store.sql")
    repository = AlbumRepository(db_connection)
    album = Album(None, "yoyo", 2020, 2)
    repository.create(album)
    assert album.id == 3
    assert repository.all() == [
        Album(1, "hello bob", 2008, 1),
        Album(2, "yoyo", 2020, 2)
    ]


def test_find_single_album(db_connection):
    db_connection.seed("seeds/record_store.sql")
    repository = AlbumRepository(db_connection)
    result = repository.find(1)
    assert result == Album(1, "Doolittle", 2008, 1)
    